Complete airport project

+++ src/controller/passenger-controller.ts
@@ -5,13 +5,32 @@ import CustomError from "../exception/custom-error";
 import { Passenger } from "../model/passenger";
 import { CreatePassengerDto } from "../dto/create-passenger-dto";
 import { UpdatePassengerDto } from "../dto/update-passenger-dto";
+import { InputValidator } from "./validator/input-validator";
+import AuthenticationUserDetails from "../security/auth-user-details";
+import NotFoundException from "../exception/not-found-exception";
+import { Paginator } from "./util/paginator";
 
 class PassengerController {
+
 
     async getAll(req: Request, res: Response, next: NextFunction) {
         try {
             const filter: PassengerFilter = JSON.parse(JSON.stringify(req.query));
-            res.send(await passengerService.retrieveByFilter(filter, ((filter.page - 1) * filter.per_page)  || 0, filter.per_page || 10));
+            InputValidator.validatePassengerInputOrThrow(filter);
+            const page = isNaN(filter.page) ? 1 : filter.page;
+            const size = isNaN(filter.per_page) ? 10 : filter.per_page;
+            const users = await passengerService.retrieveByFilter(
+                filter, 
+                Paginator.getOffset(page, size), 
+                size);
+            const totalEntries: number = await passengerService.retrieveTotalEntries(filter);
+            res.send({
+                passengers: users,
+                page: page,
+                per_page: size,
+                total_entries: totalEntries as number || 0,
+                total_pages: Paginator.getTotalPages(totalEntries, size)
+            });
         } catch(err: any) {
             next(new CustomError(err.code || 500, err.message));
         }
@@ -19,6 +38,7 @@ class PassengerController {
 
     async getById(req: Request, res: Response, next: NextFunction) {
         try {
+            InputValidator.validateIntRouteParamOrThrow(req.params.passenger_id);
             const id = parseInt(req.params.passenger_id);
             res.send(await passengerService.retrieveById(id));
         } catch(err: any) {
@@ -28,9 +48,17 @@ class PassengerController {
 
     async create(req: Request, res: Response, next: NextFunction) {
         try {
+            const user: AuthenticationUserDetails = res.locals.user;
             const body: CreatePassengerDto = req.body;
+            InputValidator.validatePassengerInputOrThrow(body);
+            if (body.userId === undefined) {
+                throw new CustomError(400, "User id is required");
+            }
+            if (user.role === 'Passenger' && user.userId != body.userId) {
+                throw new NotFoundException(404, 'Not found');
+            }
             const passenger: Passenger = {...body};
-            res.status(201).send(await passengerService.create(passenger, body.user_id || null));
+            res.status(201).send(await passengerService.create(passenger, body.userId));
         } catch(err: any) {
             next(new CustomError(err.code || 500, err.message));
         }
@@ -38,8 +66,15 @@ class PassengerController {
 
     async update(req: Request, res: Response, next: NextFunction) {
         try {
-            const body: UpdatePassengerDto = req.body;
+            InputValidator.validateIntRouteParamOrThrow(req.params.passenger_id);
             const id = parseInt(req.params.passenger_id);
+            const body: UpdatePassengerDto = req.body;
+            const passenger = await passengerService.retrieveById(id);
+            const user: AuthenticationUserDetails = res.locals.user;
+            if (user.role === 'Passenger' && passenger.userId != user.userId) {
+                throw new NotFoundException(404, 'Not found');
+            }
+            InputValidator.validatePassengerInputOrThrow(body);
             const newData: Passenger = {...body};
             res.send(await passengerService.update(newData, id));
         } catch(err: any) {
@@ -49,7 +84,13 @@ class PassengerController {
 
     async deleteById(req: Request, res: Response, next: NextFunction) {
         try {
+            InputValidator.validateIntRouteParamOrThrow(req.params.passenger_id);
             const id = parseInt(req.params.passenger_id);
+            const passenger = await passengerService.retrieveById(id);
+            const user: AuthenticationUserDetails = res.locals.user;
+            if (user.role === 'Passenger' && passenger.userId !== user.userId) {
+                throw new NotFoundException(404, 'Not found');
+            }
             await passengerService.delete(id);
             res.sendStatus(204);
         } catch(err: any) {
